Moving libs/common to libs/help
